Programming to an interface, not an implementation
Since the quack and fly methods were deemed to be the most likely to
change, they were removed from the concrete Duck implementation.

The behaviours were made into an interface (the behaviours are in
essence less likely to change, only the details of the interface are).

These behaviours are decided during runtime, plus other modules can make
use of the behaviours should they wish.